Move tests for IPython.paths
